+++ src/tad/DI52/Bindings/Resolver.php
@@ -58,6 +58,11 @@ class tad_DI52_Bindings_Resolver implements tad_DI52_Bindings_ResolverInterface
     protected $resolvingDecorator = false;
 
     /**
+     * @var array
+     */
+    protected $customBindings = array();
+
+    /**
      * @var tad_DI52_Container
      */
     private $container;
@@ -226,6 +231,24 @@ class tad_DI52_Bindings_Resolver implements tad_DI52_Bindings_ResolverInterface
         $this->decoratorsChain[$classOrInterface] = $decorators;
     }
 
+    /**
+     * Binds a class or interface implementation to a specific class resolution.
+     * When resolving `customClass` requests for the `classOrInterface` will be bound to `implementation`.
+     *
+     * @param string $customClass
+     * @param string $classOrInterface
+     * @param mixed $implementation
+     *
+     * @return mixed
+     */
+    public function bindFor($customClass, $classOrInterface, $implementation)
+    {
+        if (empty($this->customBindings[$customClass])) {
+            $this->customBindings[$customClass] = array();
+        }
+        $this->customBindings[$customClass][$classOrInterface] = $implementation;
+    }
+
     protected function bootServiceProvider(tad_DI52_ServiceProviderInterface $serviceProvider)
     {
         return $serviceProvider->boot();
@@ -309,7 +332,8 @@ class tad_DI52_Bindings_Resolver implements tad_DI52_Bindings_ResolverInterface
             $serviceProvider->register();
         }
 
-        $isBound = isset($this->bindings[$classOrInterface]);
+        $isCustomBound = array_key_exists($classOrInterface, $this->customBindings);
+        $isBound = $isCustomBound || isset($this->bindings[$classOrInterface]);
         $isDecoratorChain = !$this->resolvingDecorator && isset($this->decoratorsChain[$classOrInterface]);
 
         $isSingleton = in_array($classOrInterface, $this->singletons);
@@ -322,7 +346,17 @@ class tad_DI52_Bindings_Resolver implements tad_DI52_Bindings_ResolverInterface
             return $this->resolvedSingletons[$classOrInterface];
         }
 
-        $resolved = $isDecoratorChain ? $this->resolveBoundDecoratorChain($classOrInterface) : $this->resolveBound($classOrInterface);
+        if ($isCustomBound) {
+            $customImplementations = $this->customBindings[$classOrInterface];
+            $subResolver = clone $this;
+            $subResolver->resetCustomBindings();
+            foreach ($customImplementations as $_classOrInterface => $_implementation) {
+                $subResolver->bind($_classOrInterface, $_implementation);
+            }
+            $resolved = $subResolver->resolve($classOrInterface);
+        } else {
+            $resolved = $isDecoratorChain ? $this->resolveBoundDecoratorChain($classOrInterface) : $this->resolveBound($classOrInterface);
+        }
 
         if ($isSingleton) {
             $this->resolvedSingletons[$classOrInterface] = $resolved;
@@ -391,7 +425,7 @@ class tad_DI52_Bindings_Resolver implements tad_DI52_Bindings_ResolverInterface
     private function resolveBound($classOrInterface)
     {
         $implementation = $this->bindings[$classOrInterface];
-        return $implementation->instance();
+        return $implementation->getImplementation() === $classOrInterface ? $this->resolveUnbound($classOrInterface) : $implementation->instance();
     }
 
     protected function resolveBoundDecoratorChain($classOrInterface)
@@ -411,5 +445,10 @@ class tad_DI52_Bindings_Resolver implements tad_DI52_Bindings_ResolverInterface
         $this->resolvingDecorator = false;
 
         return $resolvedDecorator;
+    }
+
+    protected function resetCustomBindings()
+    {
+        $this->customBindings = array();
     }
 }