+++ tests/data/CustomClassOneExtension.php
@@ -1,0 +1,6 @@
+<?php
+
+class CustomClassOneExtension extends ClassOne
+{
+
+}