+++ changelog.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- support for the custom bindings
+- support for same class singleton binding
 
 ## [1.2.5] - 2016-03-06
 ### Added